image_downloader/requirements.txt
aiohttp==3.8.1
fake-useragent==0.1.11
pillow==9.2.0
redis==4.3.4
requests==2.28.1
